% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{Test Function}
\usage{
fn(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn("testing")

\testonly{testex::testex(
  . == "testing 1 2 3",
  startsWith(., "testing")
)}

\testonly{testex::testthat_expect(
  expect_equal("testing 1 2 3")
)}

\testonly{testex::testthat_block(
  test_that("fn gives expected results", {
    expect_equal(., "testing 1 2 3")
  })
)}

}
