% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen}
\alias{fn_roxygen}
\title{Test Function}
\usage{
fn_roxygen(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
\dontshow{
  value <- "testing"
}

fn_roxygen(value)\testonly{
testex::testex(
  identical(., "testing 1 2 3"),
  example = "fn.R:30:35"
)}

\dontrun{
  stop("this won't work")
}

fn_roxygen("testing")\testonly{
testex::testex(
  grepl("\\\\d", .),
  startsWith(., "testing"),
  example = "fn.R:37:42"
)}

fn_roxygen("testing")\testonly{
testex::testex(
  identical(., {
      "testing 1 2 3"
  }),
  example = "fn.R:45:46"
)}

fn_roxygen("testing")
# untested trailing example
}
