% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{Test Function}
\usage{
fn(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example showcases how you might write "raw" tests within your examples.
You could use \verb{\testonly} directly, or use \code{testex::testex()} to use
\code{.}-syntax.
}
\examples{
fn("testing")

\testonly{testex::testex(
  . == "testing 1 2 3",
  startsWith(., "testing")
)}

\testonly{testex::testthat_block(
  test_that("fn gives expected results", {
    expect_equal(., "testing 1 2 3")
  })
)}

}
