% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen}
\alias{fn_roxygen}
\title{Test Function}
\usage{
fn_roxygen(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example introduces the \verb{@expect} tag, either a value or an expression
using the \code{.}-syntax to test the last example result.
}
\examples{
\dontshow{
  value <- "testing"
}

fn_roxygen(value)\testonly{
testex::testex(
  identical(., "testing 1 2 3"),
  example = "fn.R:41:46"
)}

\dontrun{
  stop("this won't work")
}

fn_roxygen("testing")\testonly{
testex::testex(
  grepl("\\\\d", .),
  startsWith(., "testing"),
  example = "fn.R:48:53"
)}

fn_roxygen("testing")\testonly{
testex::testex(
  identical(., {
      "testing 1 2 3"
  }),
  example = "fn.R:56:57"
)}

fn_roxygen("testing")
# untested trailing example
}
