% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen}
\alias{fn_roxygen}
\title{Test Function}
\usage{
fn_roxygen(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen("testing") \testonly{
testex::testex(
  identical(., "testing 1 2 3"),
  source = "fn.R:34:35"
)}

fn_roxygen("testing") \testonly{
testex::testex(
  grepl("\\\\d", .),
  startsWith(., "tasting"),
  source = "fn.R:37:38"
)}

fn_roxygen("testing") \testonly{
testex::testex(
  identical(., {
      "testing 1 2 3"
  }),
  source = "fn.R:41:42"
)}

fn_roxygen("testing")
# untested trailing example
}
