% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen}
\alias{fn_roxygen}
\title{Test Function}
\usage{
fn_roxygen(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen("testing") \testonly{
testex::testex(
  identical(., "testing 1 2 3")
)}

fn_roxygen("testing") \testonly{
testex::testex(
  grepl("\\\\d", .),
  startsWith(., "testing")
)}

fn_roxygen("testing") \testonly{
testex::testex(
  identical(., {
      "testing 1 2 3"
  })
)}

fn_roxygen("testing")
# untested trailing example
}
