% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen}
\alias{fn_roxygen}
\title{Test Function}
\usage{
fn_roxygen(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen("testing") \testonly{
testex::testex(
  identical(., "testing 1 2 3"),
  obj    = "fn_roxygen",
  source = "fn.R:34:35",
  tests  = "fn.R:36:36"
)}

fn_roxygen("testing") \testonly{
testex::testex(
  grepl("\\\\d", .),
  startsWith(., "tasting"),
  obj    = "fn_roxygen",
  source = "fn.R:37:38",
  tests  = "fn.R:39:40"
)}

fn_roxygen("testing") \testonly{
testex::testex(
  identical(., {
      "testing 1 2 3"
  }),
  obj    = "fn_roxygen",
  source = "fn.R:41:42",
  tests  = "fn.R:43:45"
)}

fn_roxygen("testing")
# untested trailing example
}
