% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_multiple}
\alias{fn_roxygen_multiple}
\alias{fn_roxygen_multiple1}
\alias{fn_roxygen_multiple2}
\title{Test Topic Covering Multiple Functions}
\usage{
fn_roxygen_multiple1(x)

fn_roxygen_multiple2(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example composes an examples section from multiple blocks.
}
\section{Functions}{
\itemize{
\item \code{fn_roxygen_multiple1()}: Ensure multiple objects' examples are combined into a single topic

\item \code{fn_roxygen_multiple2()}: Ensure multiple objects' examples are combined into a single topic

}}
\examples{
fn_roxygen_multiple1("testing")
\testonly{testex::testex(grepl("\\\\d", .))}
\testonly{testex::testex(startsWith(., "testing"))}
fn_roxygen_multiple2("testing")
\testonly{testex::testex(grepl("\\\\d", .))}
\testonly{testex::testex(startsWith(., "testing"))}
}
