% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen("testing") \testonly{
testex::testthat_block(test_that("fn.R [62:63]", {
  expect_equal(., "testing 1 2 3")
  expect_match(., "^testing")
}))}

fn_roxygen("testing") \testonly{
testex::testthat_block(test_that("fn.R [66:67]", {
  expect_equal(., "testing 1 2 3")
  expect_match(., "^testing")
}))}
}
