% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example introduces \code{testthat}-style tests using in-line \verb{@testthat}
roxygen tags.
}
\examples{
fn_roxygen_testthat("testing")\testonly{
testex::testthat_block(test_that("example tests at `fn.R:81:82`", {
  testex::with_srcref("fn.R:83:83", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:84:84", expect_match(., "^testing"))
}),
  example = "fn.R:81:82"
)}

fn_roxygen_testthat("testing")\testonly{
testex::testthat_block(test_that("example tests at `fn.R:85:86`", {
  testex::with_srcref("fn.R:87:87", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:88:88", expect_match(., "^testing"))
}),
  example = "fn.R:85:86"
)}

}
