% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen_testthat("testing") \testonly{
testex::testthat_block(test_that("`fn_roxygen_testthat` example tests (fn.R:57:58)", {
  testex::with_srcref("fn.R:59:59", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:60:60", expect_match(., "^testing"))
}),
  obj     = "fn_roxygen_testthat",
  example = "fn.R:57:58"
)}

fn_roxygen_testthat("testing") \testonly{
testex::testthat_block(test_that("`fn_roxygen_testthat` example tests (fn.R:61:62)", {
  testex::with_srcref("fn.R:63:63", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:64:64", expect_match(., "^testing"))
}),
  obj     = "fn_roxygen_testthat",
  example = "fn.R:61:62"
)}
}
