% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\description{
Test Function
}
\examples{
fn_roxygen_testthat("testing")\testonly{
testex::testthat_block(test_that("example tests at `fn.R:65:66`", {
  testex::with_srcref("fn.R:67:67", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:68:68", expect_match(., "^testing"))
}),
  example = "fn.R:65:66"
)}

fn_roxygen_testthat("testing")\testonly{
testex::testthat_block(test_that("example tests at `fn.R:69:70`", {
  testex::with_srcref("fn.R:71:71", expect_equal(., "testing 1 2 3"))
  testex::with_srcref("fn.R:72:72", expect_match(., "^testing"))
}),
  example = "fn.R:69:70"
)}

}
