% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example introduces \code{testthat}-style tests using in-line \verb{@testthat}
\code{roxygen2} tags.
}
\examples{
fn_roxygen_testthat("testing")
\testonly{
testex::testex(tag = "testthat", srcref = "fn.R:83:83", 
expect_equal(., "testing 1 2 3"))}
\testonly{
testex::testex(tag = "testthat", srcref = "fn.R:84:84", 
expect_match(., "^testing"))}
fn_roxygen_testthat("testing")
\testonly{
testex::testex(tag = "testthat", srcref = "fn.R:87:87", 
expect_equal(., "testing 1 2 3"))}
\testonly{
testex::testex(tag = "testthat", srcref = "fn.R:88:88", 
expect_match(., "^testing"))}
}
