% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_roxygen_testthat}
\alias{fn_roxygen_testthat}
\title{Test Function}
\usage{
fn_roxygen_testthat(x)
}
\arguments{
\item{x}{A thing}
}
\value{
The pasted thing
}
\description{
This example introduces \code{testthat}-style tests using in-line \verb{@testthat}
roxygen tags.
}
\examples{
fn_roxygen_testthat("testing")
\testonly{
testex::testthat_block(test_that("example tests at `fn.R:83:83`", {
testex::with_srcref("fn.R:83:83", expect_equal(., "testing 1 2 3"))
}))
}
\testonly{
testex::testthat_block(test_that("example tests at `fn.R:84:84`", {
testex::with_srcref("fn.R:84:84", expect_match(., "^testing"))
}))
}
fn_roxygen_testthat("testing")
\testonly{
testex::testthat_block(test_that("example tests at `fn.R:88:88`", {
testex::with_srcref("fn.R:88:88", expect_equal(., "testing 1 2 3"))
}))
}
\testonly{
testex::testthat_block(test_that("example tests at `fn.R:89:89`", {
testex::with_srcref("fn.R:89:89", expect_match(., "^testing"))
}))
}
}
